Issue #66 first draft
